+++ tsMuxer/metaDemuxer.cpp
@@ -207,7 +207,7 @@ void METADemuxer::openFile(const string& streamName)
     }
 
     H264StreamReader::SeiMethod primarySEI = H264StreamReader::SEI_NotDefined;
-    for (int i = 0; i < m_codecInfo.size(); ++i)
+    for (size_t i = 0; i < m_codecInfo.size(); ++i)
     {
         H264StreamReader* reader = dynamic_cast<H264StreamReader*>(m_codecInfo[i].m_streamReader);
         if (reader && !reader->isSubStream())
@@ -220,7 +220,7 @@ void METADemuxer::openFile(const string& streamName)
     bool warned = false;
     if (primarySEI != H264StreamReader::SEI_NotDefined)
     {
-        for (int i = 0; i < m_codecInfo.size(); ++i)
+        for (size_t i = 0; i < m_codecInfo.size(); ++i)
         {
             H264StreamReader* reader = dynamic_cast<H264StreamReader*>(m_codecInfo[i].m_streamReader);
             if (reader && reader->isSubStream())
@@ -287,10 +287,10 @@ int METADemuxer::addPGSubStream(const string& codec, const string& _codecStreamN
 std::vector<MPLSPlayItem> METADemuxer::mergePlayItems(const std::vector<MPLSParser>& mplsInfoList)
 {
     std::vector<MPLSPlayItem> result;
-    for (int i = 0; i < mplsInfoList.size(); ++i)
+    for (size_t i = 0; i < mplsInfoList.size(); ++i)
     {
         const MPLSParser& mplsInfo = mplsInfoList[i];
-        for (int j = 0; j < mplsInfo.m_playItems.size(); ++j)
+        for (size_t j = 0; j < mplsInfo.m_playItems.size(); ++j)
         {
             result.push_back(mplsInfo.m_playItems[j]);
         }
@@ -326,11 +326,11 @@ int METADemuxer::addStream(const string codec, const string& codecStreamName, co
     {
         mplsInfoList = getMplsInfo(codecStreamName);
         vector<string> mplsNames = splitQuotedStr(codecStreamName.c_str(), '+');
-        for (int k = 0; k < mplsInfoList.size(); ++k)
+        for (size_t k = 0; k < mplsInfoList.size(); ++k)
         {
             MPLSParser& mplsInfo = mplsInfoList[k];
             string unquotedStreamName = unquoteStr(mplsNames[k]);
-            for (int i = 0; i < mplsInfo.m_playItems.size(); ++i)
+            for (size_t i = 0; i < mplsInfo.m_playItems.size(); ++i)
             {
                 string playItemName;
                 if (isSubStream)
@@ -1129,7 +1129,7 @@ const std::vector<MPLSParser> METADemuxer::getMplsInfo(const string& mplsFileNam
     std::vector<MPLSParser> result;
 
     std::vector<std::string> mplsFiles = splitQuotedStr(mplsFileName.c_str(), '+');
-    for (int i = 0; i < mplsFiles.size(); ++i)
+    for (size_t i = 0; i < mplsFiles.size(); ++i)
     {
         MPLSCache::iterator itr = m_mplsStreamMap.find(mplsFiles[i]);
         if (itr != m_mplsStreamMap.end())
